packages/pollination-honeybee-radiance/pollination-honeybee-radiance-0.7.1.tar.gz/pollination-honeybee-radiance-0.7.1/tests/translate_test.py
from pollination.honeybee_radiance.translate import CreateRadianceFolderGrid, \
    CreateRadianceFolderView, CreateRadiantEnclosureInfo

from queenbee.plugin.function import Function


def test_create_radiance_folder_grid():
    function = CreateRadianceFolderGrid().queenbee
    assert function.name == 'create-radiance-folder-grid'
    assert isinstance(function, Function)


def test_create_radiance_folder_view():
    function = CreateRadianceFolderView().queenbee
    assert function.name == 'create-radiance-folder-view'
    assert isinstance(function, Function)


def test_create_radiant_enclosure_info():
    function = CreateRadiantEnclosureInfo().queenbee
    assert function.name == 'create-radiant-enclosure-info'
    assert isinstance(function, Function)
